requirements.txt
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
tqdm==4.65.0
scikit-learn==1.2.2
pybind11==2.10.4
pybind11[global]==2.10.4
pytest==7.3.0
flwr==1.3.0
grad-cam==1.4.6
timm==0.6.13
torch==1.12.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
mpi4py==3.1.4
tqdm
comet_ml
python-dotenv
# ipykernel==6.22.0
# notebook==6.5.4
